Issue #35: moved TemplateChecker to new package — com.cdd.bao.validator.
